requirements.txt
gradio
llama_cpp_python